refactor(broker): extract feature-gated BrokerBuilder into test_utils.rs
Moves the inline `pub mod test_utils { ... }` block (BrokerBuilder + make_any_provider) out of lib.rs into its own file. Still gated on the `test-utils` feature so it stays out of release builds.

lib.rs replaces the ~135-line inline module with a one-line declaration: `#[cfg(feature = "test-utils")] pub mod test_utils;`.
